+++ Boot.lua
@@ -1,0 +1,9 @@
+---@class AE_Addon
+local addon = select(2, ...)
+
+addon.Libs = addon.Libs or {}
+addon.Libs.LibDataBroker = LibStub("LibDataBroker-1.1")
+addon.Libs.LibDBIcon = LibStub("LibDBIcon-1.0")
+addon.Libs.AceDB = LibStub("AceDB-3.0")
+addon.Libs.AceAddon = LibStub("AceAddon-3.0")
+addon.Libs.LiqUI = LibStub("LiqUI-1.0")