+++ Data/Keystones.lua
@@ -1,0 +1,16 @@
+---@type string
+local addonName = select(1, ...)
+---@class AE_Addon
+local addon = select(2, ...)
+
+---@class AE_Data
+local Data = addon.Data
+
+---@type AE_Keystone[]
+Data.keystones = {
+  {seasonID = 11, seasonDisplayID = 3, itemID = 151086},
+  {seasonID = 12, seasonDisplayID = 4, itemID = 180653},
+  {seasonID = 13, seasonDisplayID = 1, itemID = 180653},
+  {seasonID = 14, seasonDisplayID = 2, itemID = 180653},
+  {seasonID = 15, seasonDisplayID = 3, itemID = 180653},
+}