+++ Debug/Types/Addon.lua
@@ -1,0 +1,26 @@
+---@class AE_Libs
+---@field AceAddon AceAddon
+---@field AceDB AceDB-3.0
+---@field LibDBIcon unknown
+---@field LibDataBroker unknown
+---@field LiqUI LiqUI
+
+---@class AE_Core : AceAddon
+
+---@class AE_Module_Main : AceModule
+---@class AE_Module_Equipment : AceModule
+---@class AE_Module_WeeklyAffixes : AceModule
+
+---@class AE_Helpers
+
+---@class AE_Addon
+---@field Core AE_Core
+---@field Libs AE_Libs
+---@field LiqUI LiqUI_Instance
+---@field Constants AE_Constants
+---@field Helpers AE_Helpers
+---@field Events AE_Events
+---@field Data AE_Data
+---@field Module_Main AE_Module_Main
+---@field Module_WeeklyAffixes AE_Module_WeeklyAffixes
+---@field Module_Equipment AE_Module_Equipment