+++ Types/Utils.lua
@@ -1,7 +1,0 @@
----@class AE_ScrollFrameConfig
----@field name string?
----@field scrollSpeedHorizontal number?
----@field scrollSpeedVertical number?
-
----@class AE_Utils
----@field ScrollCollection table